bug: price 단위 int -> long으로 수정 2

+++ src/main/java/_team/earnedit/entity/Wish.java
@@ -55,7 +55,7 @@ public class Wish {
     private String url;
 
 
-    public void update(String name, int price, String itemImage, String vendor, String url) {
+    public void update(String name, long price, String itemImage, String vendor, String url) {
         this.name = name;
         this.price = price;
         this.itemImage = itemImage;
